# src/main/java/com/rozetka/dto/response/AddressResponseDTO.java
package com.rozetka.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class AddressResponseDTO {

    private Long id;
    private String street;
    private String city;
    private String country;
}