# src/main/java/com/rozetka/dto/response/ProductCategoryResponseDTO.java
package com.rozetka.dto.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ProductCategoryResponseDTO {

    private Long id;
    private String categoryName;
}